requirements.txt
# core
clearml==2.1.5
clearml-agent==2.0.7

# ml
scikit-learn==1.4.2
numpy==1.26.4
pandas==2.2.2
joblib==1.4.2

# dataset
datasets==2.19.0
huggingface-hub==0.23.0

# config
pyyaml==6.0.1
python-dotenv==1.0.1

# visualization
matplotlib==3.8.4
Pillow==10.3.0
